pyproject.toml
[project]
name = "language-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.5.2",
    "optuna>=4.4.0",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "torch",
    "torchaudio",
    "torchvision",
]
